remove review tab and create affiliations tab
